requirements.txt
disnake>=2.8.1
python-dotenv>=1.0.0
asyncio>=3.4.3
redis>=4.5.5